Simplify the main loop by avoiding Streams
I'm pretty hopeful that dropping the Future that was not chosen by
select() doesn't lose any events, because its poll() was either (a)
never called, or (b) returned Pending.  If it really could've done
something, then it will be immediately ready when it's recreated on the
next iteration.

Or at least that's the story I'm sticking to for now...